Added ability use decimals value for ephemeris steps in AstroCalc/Ephemeris tool (fix #1289)

+++ src/gui/AstroCalcCustomStepsDialog.hpp
@@ -39,7 +39,7 @@ public slots:
 	void setVisible(bool);
 
 	void populateUnitMeasurementsList();
-	void saveTimeStep(int value);
+	void saveTimeStep(double value);
 	void saveUnitMeasurement(int index);
 
 private:
